cli/requirements.txt
click>=8.0.0
requests>=2.25.0
pyyaml>=5.4.0 